requirements.txt
# Основные зависимости проекта
python-telegram-bot==20.7
python-docx==1.1.0
requests==2.31.0
python-dotenv==1.0.0
qdrant-client==1.15.1
openai==1.12.0
# Удалено: google-genai - больше не используется
beautifulsoup4==4.12.2
# Временно отключено из-за проблем с Python 3.13
# lxml==4.9.3
schedule==1.2.0
httpx==0.25.2
asyncio-throttle==1.0.2
datetime==5.4
# Временно отключены из-за проблем совместимости с Python 3.13
# pandas==2.1.4
# numpy==1.25.2

# Временно отключены астрологические расчеты
# pyephem==4.1.5
# swisseph==2.10.03.2

# Для работы с базами данных
sqlalchemy==1.4.50
alembic==1.13.1

# Утилиты
loguru==0.7.2
# Временно отключено из-за проблем с Python 3.13
# pydantic==2.5.3
